remove viewlet ts and ipanel
