+++ src/vs/workbench/browser/panecomposite.ts
@@ -3,21 +3,23 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { Registry } from 'vs/platform/registry/common/platform';
+import { Composite, CompositeDescriptor, CompositeRegistry } from 'vs/workbench/browser/composite';
+import { IConstructorSignature0, BrandedService, IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
+import { URI } from 'vs/base/common/uri';
 import { Dimension } from 'vs/base/browser/dom';
-import { IContextMenuService } from 'vs/platform/contextview/browser/contextView';
-import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
-import { IThemeService } from 'vs/platform/theme/common/themeService';
-import { IView } from 'vs/workbench/common/views';
-import { IStorageService } from 'vs/platform/storage/common/storage';
-import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
-import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
-import { Composite } from 'vs/workbench/browser/composite';
-import { IWorkspaceContextService } from 'vs/platform/workspace/common/workspace';
-import { ViewPaneContainer, ViewsSubMenu } from './parts/views/viewPaneContainer';
-import { IPaneComposite } from 'vs/workbench/common/panecomposite';
+import { IActionViewItem } from 'vs/base/browser/ui/actionbar/actionbar';
 import { IAction, Separator } from 'vs/base/common/actions';
 import { SubmenuItemAction } from 'vs/platform/actions/common/actions';
-import { IActionViewItem } from 'vs/base/browser/ui/actionbar/actionbar';
+import { IContextMenuService } from 'vs/platform/contextview/browser/contextView';
+import { IStorageService } from 'vs/platform/storage/common/storage';
+import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
+import { IThemeService } from 'vs/platform/theme/common/themeService';
+import { IWorkspaceContextService } from 'vs/platform/workspace/common/workspace';
+import { ViewPaneContainer, ViewsSubMenu } from 'vs/workbench/browser/parts/views/viewPaneContainer';
+import { IPaneComposite } from 'vs/workbench/common/panecomposite';
+import { IView } from 'vs/workbench/common/views';
+import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
 
 export abstract class PaneComposite extends Composite implements IPaneComposite {
 
@@ -134,3 +136,74 @@ export abstract class PaneComposite extends Composite implements IPaneComposite 
 
 	protected abstract createViewPaneContainer(parent: HTMLElement): ViewPaneContainer;
 }
+
+
+/**
+ * A Pane Composite descriptor is a leightweight descriptor of a Pane Composite in the workbench.
+ */
+export class PaneCompositeDescriptor extends CompositeDescriptor<PaneComposite> {
+
+	static create<Services extends BrandedService[]>(
+		ctor: { new(...services: Services): PaneComposite },
+		id: string,
+		name: string,
+		cssClass?: string,
+		order?: number,
+		requestedIndex?: number,
+		iconUrl?: URI
+	): PaneCompositeDescriptor {
+
+		return new PaneCompositeDescriptor(ctor as IConstructorSignature0<PaneComposite>, id, name, cssClass, order, requestedIndex, iconUrl);
+	}
+
+	private constructor(
+		ctor: IConstructorSignature0<PaneComposite>,
+		id: string,
+		name: string,
+		cssClass?: string,
+		order?: number,
+		requestedIndex?: number,
+		readonly iconUrl?: URI
+	) {
+		super(ctor, id, name, cssClass, order, requestedIndex);
+	}
+}
+
+export const Extensions = {
+	Viewlets: 'workbench.contributions.viewlets',
+	Panels: 'workbench.contributions.panels'
+};
+
+export class PaneCompositeRegistry extends CompositeRegistry<PaneComposite> {
+
+	/**
+	 * Registers a viewlet to the platform.
+	 */
+	registerViewlet(descriptor: PaneCompositeDescriptor): void {
+		super.registerComposite(descriptor);
+	}
+
+	/**
+	 * Deregisters a viewlet to the platform.
+	 */
+	deregisterViewlet(id: string): void {
+		super.deregisterComposite(id);
+	}
+
+	/**
+	 * Returns the viewlet descriptor for the given id or null if none.
+	 */
+	getViewlet(id: string): PaneCompositeDescriptor {
+		return this.getComposite(id) as PaneCompositeDescriptor;
+	}
+
+	/**
+	 * Returns an array of registered viewlets known to the platform.
+	 */
+	getViewlets(): PaneCompositeDescriptor[] {
+		return this.getComposites() as PaneCompositeDescriptor[];
+	}
+}
+
+Registry.add(Extensions.Viewlets, new PaneCompositeRegistry());
+Registry.add(Extensions.Panels, new PaneCompositeRegistry());
