+++ src/vs/workbench/browser/parts/sidebar/sidebarPart.ts
@@ -7,7 +7,7 @@ import 'vs/css!./media/sidebarpart';
 import 'vs/workbench/browser/parts/sidebar/sidebarActions';
 import { Registry } from 'vs/platform/registry/common/platform';
 import { CompositePart } from 'vs/workbench/browser/parts/compositePart';
-import { PaneCompositeRegistry, Extensions as ViewletExtensions, PaneCompositeDescriptor } from 'vs/workbench/browser/viewlet';
+import { PaneCompositeRegistry, Extensions as ViewletExtensions, PaneCompositeDescriptor, PaneComposite } from 'vs/workbench/browser/panecomposite';
 import { IViewletService } from 'vs/workbench/services/viewlet/browser/viewlet';
 import { IWorkbenchLayoutService, Parts, Position as SideBarPosition } from 'vs/workbench/services/layout/browser/layoutService';
 import { SidebarFocusContext, ActiveViewletContext } from 'vs/workbench/common/viewlet';
@@ -33,7 +33,6 @@ import { CompositeDragAndDropObserver } from 'vs/workbench/browser/dnd';
 import { IViewDescriptorService, ViewContainerLocation } from 'vs/workbench/common/views';
 import { Gesture, EventType as GestureEventType } from 'vs/base/browser/touch';
 import { IPaneComposite } from 'vs/workbench/common/panecomposite';
-import { PaneComposite } from 'vs/workbench/browser/panecomposite';
 
 export class SidebarPart extends CompositePart<PaneComposite> implements IViewletService {
