+++ src/vs/workbench/common/panel.ts
@@ -4,7 +4,6 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { localize } from 'vs/nls';
-import { IComposite } from 'vs/workbench/common/composite';
 import { RawContextKey } from 'vs/platform/contextkey/common/contextkey';
 
 export const ActivePanelContext = new RawContextKey<string>('activePanel', '', localize('activePanel', "The identifier of the active panel"));
@@ -12,5 +11,3 @@ export const PanelFocusContext = new RawContextKey<boolean>('panelFocus', false,
 export const PanelPositionContext = new RawContextKey<string>('panelPosition', 'bottom', localize('panelPosition', "The position of the panel, either 'left', 'right' or 'bottom'"));
 export const PanelVisibleContext = new RawContextKey<boolean>('panelVisible', false, localize('panelVisible', "Whether the panel is visible"));
 export const PanelMaximizedContext = new RawContextKey<boolean>('panelMaximized', false, localize('panelMaximized', "Whether the panel is maximized"));
-
-export interface IPanel extends IComposite { }
