+++ src/vs/workbench/services/panel/common/panelService.ts
@@ -4,11 +4,11 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { Event } from 'vs/base/common/event';
-import { IPanel } from 'vs/workbench/common/panel';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { IBadge } from 'vs/workbench/services/activity/common/activity';
 import { IDisposable } from 'vs/base/common/lifecycle';
 import { IProgressIndicator } from 'vs/platform/progress/common/progress';
+import { IPaneComposite } from 'vs/workbench/common/panecomposite';
 
 export const IPanelService = createDecorator<IPanelService>('panelService');
 
@@ -22,18 +22,18 @@ export interface IPanelService {
 
 	readonly _serviceBrand: undefined;
 
-	readonly onDidPanelOpen: Event<{ readonly panel: IPanel, readonly focus: boolean }>;
-	readonly onDidPanelClose: Event<IPanel>;
+	readonly onDidPanelOpen: Event<{ readonly panel: IPaneComposite, readonly focus: boolean }>;
+	readonly onDidPanelClose: Event<IPaneComposite>;
 
 	/**
 	 * Opens a panel with the given identifier and pass keyboard focus to it if specified.
 	 */
-	openPanel(id?: string, focus?: boolean): Promise<IPanel | undefined>;
+	openPanel(id?: string, focus?: boolean): Promise<IPaneComposite | undefined>;
 
 	/**
 	 * Returns the current active panel or null if none
 	 */
-	getActivePanel(): IPanel | undefined;
+	getActivePanel(): IPaneComposite | undefined;
 
 	/**
 	 * Returns the panel by id.
